- Changed Task Profiler to be a little less greedy for memory. - Added task name info functions to several task object.

+++ src/lib/world/taskfn.h
@@ -451,7 +451,7 @@ namespace madness {
             return detail::get_mem_func_ptr(wrapper);
         }
 
-        virtual void get_id(std::pair<void*,unsigned long>& id) const {
+        virtual void get_id(std::pair<void*,unsigned short>& id) const {
             return make_id(id, get_func(func_));
         }
